requirements.txt
pandas==2.1.1
scikit-learn==1.3.1
sagemaker~=2.187.0
numpy~=1.26.0
joblib~=1.3.2
boto3~=1.28.52
pyelftools~=0.30
capstone~=5.0.1
requests~=2.31.0
mitmproxy~=10.1.5
pefile~=2023.2.7
matplotlib~=3.8.2
seaborn~=0.13.0